initial commit for edward-bot
i've written some code for an "edward bot". it will eventually receive
and reply to gpg-encrypted emails.

some segments of code are based on work by Josh Drake and Lisa Maginnis.

this commit contains a large amount of code for a first commit. while it
was my intention to start using vcs once i had useful code, it's now
well past that time.