Variable name changes. Removed gpg sig. analyzer
Many variable names were changed to make the code easier to understand.

Signature analysis was removed and replaced with collection of keys used
in gpg signing. This will be useful for choosing the public key to use
when encrypting the reply message.

These two commits should have been separate but the edits got mixed
together.