changed message generation logic; some refactoring
now edward remains silent about not receiving a public key, unless we
are not given a key and need one for verifying a signature. some
refactoring was required.

edward now complains to the user about lack of of encryption and
signatures in received emails.

i also changed the handling of newlines in email responses.