check to see if a reply encryption key can encrypt
sometimes keys are revoked or expire. if so, they should not be
encrypted to. edward now checks to make sure a key is good before
encrypting a reply with it. if only bad keys are available, edward
complains. however the complaint edward makes is that he does not have
the user's public key. this is because there is currently no translation
set for a more accurate message.

some other refactoring has been performed as well.