verify signatures even if key is new to edward
previously when users sent edward their gpg key and signed that message
with that key, users would not receive a message noting signature
verification success.

this commit processes emails twice so that keys are imported the first
time and signatures are verified the second time around. it is not the
most elegant method, but it works and lengthens run time very little.
for now debug output mentions imported keys twice but debug output is
not visible to users.