add "From: " statement to email header
It is automatically inserted at the bottom of the header by python's
SMTP library, but this appears to be not good enough. Some SMTP servers
might be cutting out the From: address if it is too low in the header.
I've placed it higher in the header, near the "To: " statement.